2007-06-22  George Giolfan  <georgegiolfan@yahoo.com>
	* Grid.cs: Fixed type in previous change.

+++ george/Source/PresentationFramework/System.Windows.Controls/Grid.cs
@@ -405,11 +405,11 @@ namespace System.Windows.Controls {
 		}
 
 		static bool ValidateRowColumn(object value) {
-			return (double)value >= 0;
+			return (int)value >= 0;
 		}
 
 		static bool ValidateSpan(object value) {
-			return (double)value > 0;
+			return (int)value > 0;
 		}
 		#endregion
